requirements.txt
websocket-client>=1.0
requests>=2.25
pygments>=2.10
configparser>=5.0